Create 3-course-tables
HTML Tables
Great job! In this lesson, you learned how to create a table, add data to it, and section the table into smaller parts that make it easier to read.

Let's review what you've learned so far:

The <table> element creates a table.
The <tr> element adds rows to a table.
To add data to a row, you can use the <td> element.
Table headings clarify the meaning of data. Headings are added with the <th> element.
Table data can span columns using the colspan attribute.
Table data can span rows using the rowspan attribute.
Tables can be split into three main sections: a head, a body, and a footer.
A table's head is created with the <thead> element.
A table's body is created with the <tbody> element.
A table's footer is created with the <tfoot> element.
All the CSS properties you learned about in this course can be applied to tables and their data.
Congratulations on completing HTML Tables!